heap allocation testing

+++ kernel/arch/i386/kernel/tty.c
@@ -15,7 +15,8 @@ static size_t terminal_column;
 static uint8_t terminal_color;
 static uint16_t* terminal_buffer;
 
-void terminal_initialize(void) {
+void terminal_initialize(void)
+{
 	terminal_row = 2;
 	terminal_column = 0;
 	terminal_color = vga_entry_color(VGA_COLOR_CYAN, VGA_COLOR_BLACK);
@@ -29,17 +30,20 @@ void terminal_initialize(void) {
 	terminal_writestring("Terminal initialized\n");
 }
 
-void terminal_setcolor(uint8_t color) {
+void terminal_setcolor(uint8_t color)
+{
 	terminal_color = color;
 }
 
-void terminal_putentryat(unsigned char c, uint8_t color, size_t x, size_t y) {
+void terminal_putentryat(unsigned char c, uint8_t color, size_t x, size_t y)
+{
 	const size_t index = y * VGA_WIDTH + x;
 	terminal_buffer[index] = vga_entry(c, color);
 }
 
-void terminal_scroll() {
-	int i, k, j;
+void terminal_scroll()
+{
+	size_t i, k, j;
 	for (k = 2, j = 1; k < VGA_HEIGHT; k++, j++) {
 		for (i = 0; i < VGA_WIDTH; i++) {
 			terminal_putentryat(terminal_buffer[k * VGA_WIDTH + i], 
@@ -51,7 +55,8 @@ void terminal_scroll() {
 	terminal_row--;
 }
 
-void terminal_putchar(char c) {
+void terminal_putchar(char c)
+{
 	unsigned char uc = c;
 	if (c == '\n') {
 		terminal_column = 0;
@@ -67,11 +72,13 @@ void terminal_putchar(char c) {
 	}
 }
 
-void terminal_write(const char* data, size_t size) {
+void terminal_write(const char* data, size_t size)
+{
 	for (size_t i = 0; i < size; i++)
 		terminal_putchar(data[i]);
 }
 
-void terminal_writestring(const char* data) {
+void terminal_writestring(const char* data)
+{
 	terminal_write(data, strlen(data));
 }
